+++ package.json
@@ -31,7 +31,7 @@
     "@emotion/styled": "^11.11.0",
     "@google-cloud/storage": "^6.11.0",
     "@hookform/resolvers": "^3.1.1",
-    "@katitb2023/database": "git+https://github.com/KATITB2023/database.git#7f9edecaf4c66866e0a753a7031bb60faf7de2c4",
+    "@katitb2023/database": "git+https://github.com/KATITB2023/database.git#540d633020dd565ef799fc0c6c35ca36359614c4",
     "@next-auth/prisma-adapter": "^1.0.5",
     "@opentelemetry/api": "^1.4.1",
     "@opentelemetry/exporter-trace-otlp-http": "^0.40.0",
@@ -60,6 +60,7 @@
     "react-dom": "18.2.0",
     "react-hook-form": "^7.45.0",
     "react-icons": "^4.10.1",
+    "react-markdown": "^8.0.7",
     "sanitize-filename": "^1.6.3",
     "sharp": "^0.32.1",
     "superjson": "1.12.2",
